Buffering for ogg123. It's not as robust as it could be, but still better than nothing.
gcc 2.95.2 20000220 does _not_ compile this right with -O1 or higher.
gdb cannot properly debug multithreaded apps (it follows the parent
process even when I tell it to follow the child explicitly), so I have
not been able to figure out where the culprit is either.

This is a Unix buffer implementation. Non-Unix platforms should replace
buffer.c with a pass-through in submit_chunk that just calls devices_write.


git-svn-id: https://svn.xiph.org/trunk/vorbis-tools@1138 0101bb08-14d6-0310-b084-bc0e0c8e3800