Another patch for lazy people (like me).
Make AM_PATH_OGG, AM_PATH_VORBIS, and AM_PATH_AO each a bit smarter.
Each of the three will now also check the $prefix for the location
of their respective libraries/header files.  This means that if you're
compiling all 4 things (ao, ogg, vorbis, vorbis-tools) all with the
same $prefix (which is a pretty common case), you don't have to specify
"--with-ao-prefix=this --with-ogg-prefix=really --with-vorbis-prefix=sucks",
you can just specify a single --prefix argument and ditch all the
--with-*-prefix arguments.

However, if you do specify any of the three --with arguments, those are
searched *first* -- the $prefix is searched second.


git-svn-id: https://svn.xiph.org/trunk/vorbis-tools@1371 0101bb08-14d6-0310-b084-bc0e0c8e3800