Require libvorbis 1.3.0 or later to build.
The oggenc program uses OV_ECTL_COUPLING_SET which appreared in
libvorbis 1.3.0. We could easily conditionalize this option,
but prefer to require that the two version upgrade in sync, so
instead the configure script checks for this version, or if
pkg-config isn't available, it checks for the symbol itself.

Note that the use of AC_CHECK_DECL() probably requires a more
recent autoconf version than the 2.53 dependency we declare,
but as usual the documentation provides no guidance as to when
this macro was added.


git-svn-id: https://svn.xiph.org/trunk/vorbis-tools@17070 0101bb08-14d6-0310-b084-bc0e0c8e3800