Update to the latest version of ACX_PTHEAD.
Fixes an issue with the macro expecting pthread_t to be
convertable from an int, which isn't true on mingw where
pthread_t is a struct.

Also move the macro out of acinclude.m4 and into its own
file in the m4 subdirectory.

Thanks to Laura Michaels for reporting the issue.



git-svn-id: https://svn.xiph.org/trunk/vorbis-tools@18791 0101bb08-14d6-0310-b084-bc0e0c8e3800